Refactor tests
Now the file organisation of the tests actually makes sense.
< organisation, testing
