+++ src/test/java/nl/tudelft/controller/collision/HorWallWallHandlerTest.java
@@ -1,0 +1,75 @@
+package nl.tudelft.controller.collision;
+
+import static org.mockito.Matchers.anyFloat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import nl.tudelft.model.Game;
+import nl.tudelft.model.wall.AbstractWall;
+import nl.tudelft.model.wall.HorMovingWall;
+import org.junit.Before;
+import org.junit.Test;
+
+public class HorWallWallHandlerTest {
+
+    private DefaultPlayerInteractionMap map;
+    private Game mockedGame;
+    private HorMovingWall mockedHorWall;
+    private AbstractWall mockedWall;
+
+    /**
+     * Create the interaction map and mocks used for testing.
+     */
+    @Before
+    public void setUp() {
+        map = new DefaultPlayerInteractionMap();
+        mockedGame = mock(Game.class);
+        mockedHorWall = mock(HorMovingWall.class);
+        mockedWall = mock(AbstractWall.class);
+    }
+
+    @Test
+    public void testHorMovingWallWallCollision1() {
+        when(mockedHorWall.getSpeed()).thenReturn(1f);
+        when(mockedHorWall.getLocX()).thenReturn(1f);
+        when(mockedHorWall.getWidth()).thenReturn(1);
+
+        when(mockedWall.getLocX()).thenReturn(1f);
+        when(mockedWall.getWidth()).thenReturn(1);
+
+        map.collide(mockedGame, mockedHorWall, mockedWall);
+
+        verify(mockedHorWall, never()).setSpeed(anyFloat());
+    }
+
+    @Test
+    public void testHorMovingWallWallCollision2() {
+        when(mockedHorWall.getSpeed()).thenReturn(1f);
+        when(mockedHorWall.getLocX()).thenReturn(2f);
+        when(mockedHorWall.getWidth()).thenReturn(1);
+
+        when(mockedWall.getLocX()).thenReturn(1f);
+        when(mockedWall.getWidth()).thenReturn(1);
+
+        map.collide(mockedGame, mockedHorWall, mockedWall);
+
+        verify(mockedHorWall, times(1)).setSpeed(anyFloat());
+    }
+
+    @Test
+    public void testHorMovingWallWallCollision3() {
+        when(mockedHorWall.getSpeed()).thenReturn(1f);
+        when(mockedHorWall.getLocX()).thenReturn(1f);
+        when(mockedHorWall.getWidth()).thenReturn(1);
+
+        when(mockedWall.getLocX()).thenReturn(5f);
+        when(mockedWall.getWidth()).thenReturn(1);
+
+        map.collide(mockedGame, mockedHorWall, mockedWall);
+
+        verify(mockedHorWall, times(1)).setSpeed(anyFloat());
+    }
+}
