+++ src/test/java/nl/tudelft/controller/collision/PlayerUtilityHandlerTest.java
@@ -1,0 +1,26 @@
+package nl.tudelft.controller.collision;
+
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
+import nl.tudelft.model.Game;
+import nl.tudelft.model.Player;
+import nl.tudelft.model.pickups.utility.Utility;
+import org.junit.Test;
+
+public class PlayerUtilityHandlerTest {
+
+    @Test
+    public void testUtilPlayerCollision() {
+        DefaultPlayerInteractionMap map = new DefaultPlayerInteractionMap();
+        Game mockedGame = mock(Game.class);
+        Utility mockedUtility = mock(Utility.class);
+        Player mockedPlayer = mock(Player.class);
+
+        map.collide(mockedGame, mockedUtility, mockedPlayer);
+
+        verify(mockedUtility, times(1)).activate(any());
+    }
+}
